split queue container
