+++ CMakeLists.txt
@@ -5,32 +5,22 @@ project(lock_free_queue)
 set(CMAKE_CXX_STANDARD 14)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
-# set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wextra -pedantic -fsanitize=address -O2")
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wextra -pedantic -O2")
 set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -g")
 
-# Fetch GoogleTest
-include(FetchContent)
-FetchContent_Declare(
-  googletest
-  URL https://github.com/google/googletest/archive/03597a01ee50ed33e9dfd640b249b4be3799d395.zip
-)
-# For Windows: Prevent overriding the parent project's compiler/linker settings
-set(gtest_force_shared_crt ON CACHE BOOL "" FORCE)
-FetchContent_MakeAvailable(googletest)
+# Option to enable or disable tests
+option(BUILD_TESTS "Build tests" ON)
 
-enable_testing()
+# Include directories
+include_directories(${CMAKE_SOURCE_DIR}/include)
 
-include_directories(include)
+# Add subdirectory for tests
+if(BUILD_TESTS)
+  add_subdirectory(test)
+endif()
 
+# Main executable
 add_executable(
-  hello_test
-  test/test_queue.cpp
+  main
+  src/main.cpp
 )
-target_link_libraries(
-  hello_test
-  GTest::gtest_main
-)
-
-include(GoogleTest)
-gtest_discover_tests(hello_test)
